chore: reorganização dos arquivos
Criação de arquivo routes.py para armazenar as rotas da API, mudança da pasta db/query para db/sql.
